Updated a few notebooks
Fixed a few typos noted by students. Also changed it so that notebooks 01_02 and 01_03 are shipped without the output cells filled. Might lead to a bunch of merge conflicts for students, but maybe we have them make their own branch, commit changes, and rebase from the main branch to handle the conflicts but preserve their working version. I'll need to test this first